vrepl: improve failure handling
Tablet dialers were set to not fail fast. This was causing grpc
to redial the same tablet indefinitely eventhough it was permanently
down. This has now been changed to fail fast.

The tablet picker was only taking health errors into account, which
happens when tablets report themseleves as unhealthy. But if a tablet
is unreachable, we have to check LastError.

Also fixed some staticcheck errors.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>