vdiff: refactor MergeSort for readability
MergeSort as a function that required a VCursor that performed
scatter queries was hard to use from outside of the engine code.

In the new change, MergeSort is a Primitive that directly drives
other sub-Primitives. This greatly simplifies the vdiff code,
removing the need for a mergeSorter and a resultStreamer.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>